Move @discriminator to core
